Move registry to notify
